+++ packages/component-library/src/components/feedback-indicator/mt-banner/mt-banner.interactive.stories.ts
@@ -1,5 +1,5 @@
-import { within, userEvent } from "@storybook/testing-library";
-import { expect } from "@storybook/jest";
+import { within, userEvent } from "@storybook/test";
+import { expect } from "@storybook/test";
 
 import meta, { type MtBannerMeta, type MtBannerStory } from "./mt-banner.stories";