+++ packages/component-library/src/components/form/mt-checkbox/mt-checkbox.interactive.stories.ts
@@ -1,5 +1,5 @@
-import { within, userEvent } from "@storybook/testing-library";
-import { expect } from "@storybook/jest";
+import { within, userEvent } from "@storybook/test";
+import { expect } from "@storybook/test";
 import { waitUntilRendered } from "../../../_internal/test-helper";
 
 import meta, { type MtCheckboxMeta, type MtCheckboxStory } from "./mt-checkbox.stories";