+++ examples/nuxt-app/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Patch Changes
 
+- Updated dependencies [e0e1741]
+  - @shopware-ag/meteor-component-library@3.2.0
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [6c6678d]
   - @shopware-ag/meteor-component-library@3.1.0